Re-organize inverted_index crate.
