+++ src/redisearch_rs/inverted_index/src/codec/full.rs
