+++ src/redisearch_rs/inverted_index/src/index/mod.rs
@@ -1,0 +1,17 @@
+/*
+ * Copyright (c) 2006-Present, Redis Ltd.
+ * All rights reserved.
+ *
+ * Licensed under your choice of the Redis Source Available License 2.0
+ * (RSALv2); or (b) the Server Side Public License v1 (SSPLv1); or (c) the
+ * GNU Affero General Public License v3 (AGPLv3).
+*/
+
+mod core;
+pub mod opaque;
+mod with_entries;
+mod with_mask;
+
+pub use self::core::*;
+pub use with_entries::EntriesTrackingIndex;
+pub use with_mask::FieldMaskTrackingIndex;
