+++ src/redisearch_rs/inverted_index/src/index/opaque.rs
