+++ src/redisearch_rs/inverted_index/src/tests/reader/geo.rs
@@ -1,0 +1,70 @@
+/*
+ * Copyright (c) 2006-Present, Redis Ltd.
+ * All rights reserved.
+ *
+ * Licensed under your choice of the Redis Source Available License 2.0
+ * (RSALv2); or (b) the Server Side Public License v1 (SSPLv1); or (c) the
+ * GNU Affero General Public License v3 (AGPLv3).
+*/
+
+use std::ptr;
+
+use crate::{FilterGeoReader, IndexReader, NumericFilter, RSIndexResult};
+use ffi::{GeoDistance_GEO_DISTANCE_M, GeoFilter};
+use pretty_assertions::assert_eq;
+
+#[test]
+fn reading_filter_based_on_geo_filter() {
+    /// Implement this FFI call for this test
+    #[unsafe(no_mangle)]
+    pub extern "C" fn isWithinRadius(gf: *const GeoFilter, d: f64, distance: *mut f64) -> bool {
+        if d > unsafe { (*gf).radius } {
+            return false;
+        }
+
+        // Tests changing the distance value
+        unsafe { *distance /= 5.0 };
+
+        true
+    }
+
+    // Make an iterator with three records having different geo distances. The last record will be
+    // filtered out based on the geo distance.
+    let iter = vec![
+        RSIndexResult::numeric(5.0).doc_id(10),
+        RSIndexResult::numeric(15.0).doc_id(11),
+        RSIndexResult::numeric(25.0).doc_id(12),
+    ];
+
+    let geo_filter = GeoFilter {
+        fieldSpec: ptr::null(),
+        lat: 0.0,
+        lon: 0.0,
+        radius: 20.0,
+        unitType: GeoDistance_GEO_DISTANCE_M,
+        numericFilters: ptr::null_mut(),
+    };
+
+    let filter = NumericFilter {
+        min: 0.0,
+        max: 0.0,
+        min_inclusive: false,
+        max_inclusive: false,
+        field_spec: ptr::null(),
+        geo_filter: &geo_filter as *const _ as *const _,
+        ascending: true,
+        limit: 0,
+        offset: 0,
+    };
+
+    let mut reader = FilterGeoReader::new(&filter, iter.into_iter());
+    let mut result = RSIndexResult::numeric(0.0);
+
+    let found = reader.next_record(&mut result).unwrap();
+    assert!(found);
+    assert_eq!(result, RSIndexResult::numeric(1.0).doc_id(10));
+
+    let found = reader.next_record(&mut result).unwrap();
+    assert!(found);
+    assert_eq!(result, RSIndexResult::numeric(3.0).doc_id(11));
+}
