+++ src/redisearch_rs/inverted_index/tests/integration/codec/freqs_fields.rs
@@ -15,8 +15,6 @@ use inverted_index::{
     freqs_fields::{FreqsFields, FreqsFieldsWide},
 };
 
-mod c_mocks;
-
 #[test]
 fn test_encode_freqs_fields() {
     // Test cases for the freqs fields encoder and decoder.
