+++ src/redisearch_rs/inverted_index/tests/integration/codec/mod.rs
@@ -1,0 +1,19 @@
+/*
+ * Copyright (c) 2006-Present, Redis Ltd.
+ * All rights reserved.
+ *
+ * Licensed under your choice of the Redis Source Available License 2.0
+ * (RSALv2); or (b) the Server Side Public License v1 (SSPLv1); or (c) the
+ * GNU Affero General Public License v3 (AGPLv3).
+*/
+
+mod doc_ids_only;
+mod fields_offsets;
+mod fields_only;
+mod freqs_fields;
+mod freqs_offsets;
+mod freqs_only;
+mod full;
+mod numeric;
+mod offsets_only;
+mod raw_doc_ids_only;
