+++ src/redisearch_rs/inverted_index/tests/integration/codec/raw_doc_ids_only.rs
@@ -9,25 +9,7 @@
 
 use std::io::Cursor;
 
-use ffi::RSQueryTerm;
 use inverted_index::{Decoder, Encoder, RSIndexResult, raw_doc_ids_only::RawDocIdsOnly};
-
-#[unsafe(no_mangle)]
-pub extern "C" fn ResultMetrics_Free(metrics: *mut ffi::RSYieldableMetric) {
-    if metrics.is_null() {
-        return;
-    }
-
-    panic!(
-        "did not expect any test to set metrics, but got: {:?}",
-        unsafe { *metrics }
-    );
-}
-
-#[unsafe(no_mangle)]
-pub extern "C" fn Term_Free(_t: *mut RSQueryTerm) {
-    // The RSQueryTerm used in those tests is stack allocated so we don't need to free it.
-}
 
 #[test]
 fn test_encode_raw_doc_ids_only() {
