+++ src/redisearch_rs/inverted_index/tests/integration/controlled_cursor.rs
