+++ src/redisearch_rs/inverted_index/tests/integration/main.rs
@@ -1,0 +1,15 @@
+/*
+ * Copyright (c) 2006-Present, Redis Ltd.
+ * All rights reserved.
+ *
+ * Licensed under your choice of the Redis Source Available License 2.0
+ * (RSALv2); or (b) the Server Side Public License v1 (SSPLv1); or (c) the
+ * GNU Affero General Public License v3 (AGPLv3).
+*/
+
+pub(crate) mod c_mocks;
+
+mod codec;
+mod controlled_cursor;
+mod index;
+mod index_result;
